Also split the implementation
